Add files via upload
This project is made with machine learning and neural networks. There are train set and test set. Train set has output, that is churn of returning subscription of people; if they reject subscription 1, if no 0. Need to trained train set, and test set's outputs should be trained and predict outputs.